wip: firebase ui
